housebudgetapp/requirements.txt
wheel
tortoise-orm[asyncpg]
tortoise-orm[accel]
asynctest
